move test samples detection inside actions folder
