+++ src/main/kotlin/org/jetbrains/research/testspark/actions/llm/LLMSampleSelectorBuilder.kt
@@ -1,11 +1,19 @@
 package org.jetbrains.research.testspark.actions.llm
 
+import com.intellij.openapi.fileTypes.FileType
+import com.intellij.openapi.fileTypes.FileTypeManager
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.roots.ProjectFileIndex
+import com.intellij.openapi.roots.ProjectRootManager
+import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiJavaFile
+import com.intellij.psi.PsiManager
+import com.intellij.psi.PsiMethod
+import com.intellij.util.containers.stream
 import com.intellij.util.ui.FormBuilder
 import org.jetbrains.research.testspark.actions.template.PanelBuilder
 import org.jetbrains.research.testspark.bundles.plugin.PluginLabelsBundle
 import org.jetbrains.research.testspark.core.test.SupportedLanguage
-import org.jetbrains.research.testspark.helpers.LLMTestSampleHelper
 import java.awt.Font
 import javax.swing.ButtonGroup
 import javax.swing.JButton
@@ -25,7 +33,7 @@ class LLMSampleSelectorBuilder(private val project: Project, private val languag
     private val defaultTestName = "<html>provide manually</html>"
     private val defaultTestCode = "// provide test method code here"
     private val testNames = mutableListOf(defaultTestName)
-    private val initialTestCodes = mutableListOf(LLMTestSampleHelper.createTestSampleClass("", defaultTestCode))
+    private val initialTestCodes = mutableListOf(createTestSampleClass("", defaultTestCode))
     private val testSamplePanelFactories: MutableList<TestSamplePanelBuilder> = mutableListOf()
     private var testSamplesCode: String = ""
 
@@ -46,7 +54,7 @@ class LLMSampleSelectorBuilder(private val project: Project, private val languag
     init {
         addListeners()
 
-        LLMTestSampleHelper.collectTestSamples(project, testNames, initialTestCodes)
+        collectTestSamples(project, testNames, initialTestCodes)
     }
 
     override fun getTitlePanel(): JPanel {
@@ -129,7 +137,8 @@ class LLMSampleSelectorBuilder(private val project: Project, private val languag
         }
 
         addButton.addActionListener {
-            val testSamplePanelBuilder = TestSamplePanelBuilder(project, middlePanel, testNames, initialTestCodes, language)
+            val testSamplePanelBuilder =
+                TestSamplePanelBuilder(project, middlePanel, testNames, initialTestCodes, language)
             testSamplePanelFactories.add(testSamplePanelBuilder)
             val testSamplePanel = testSamplePanelBuilder.getTestSamplePanel()
             val codeScrollPanel = testSamplePanelBuilder.getCodeScrollPanel()
@@ -173,4 +182,55 @@ class LLMSampleSelectorBuilder(private val project: Project, private val languag
             testSamplePanelFactory.enabledComponents(isEnabled)
         }
     }
+
+    /**
+     * Retrieves a list of test samples from the given project.
+     *
+     * @return A list of strings, representing the names of the test samples.
+     */
+    private fun collectTestSamples(project: Project, testNames: MutableList<String>, initialTestCodes: MutableList<String>) {
+        val projectFileIndex: ProjectFileIndex = ProjectRootManager.getInstance(project).fileIndex
+        val javaFileType: FileType = FileTypeManager.getInstance().getFileTypeByExtension("java")
+
+        projectFileIndex.iterateContent { file ->
+            if (file.fileType === javaFileType) {
+                try {
+                    val psiJavaFile = (PsiManager.getInstance(project).findFile(file) as PsiJavaFile)
+                    val psiClass = psiJavaFile.classes[
+                        psiJavaFile.classes.stream().map { it.name }.toArray()
+                            .indexOf(psiJavaFile.name.removeSuffix(".java")),
+                    ]
+                    var imports = psiJavaFile.importList?.allImportStatements?.map { it.text }?.toList()
+                        ?.joinToString("\n") ?: ""
+                    if (psiClass.qualifiedName != null && psiClass.qualifiedName!!.contains(".")) {
+                        imports += "\nimport ${psiClass.qualifiedName?.substringBeforeLast(".") + ".*"};"
+                    }
+                    psiClass.allMethods.forEach { method ->
+                        val annotations = method.modifierList.annotations
+                        annotations.forEach { annotation ->
+                            if (annotation.qualifiedName == "org.junit.jupiter.api.Test" || annotation.qualifiedName == "org.junit.Test") {
+                                val code: String = createTestSampleClass(imports, method.text)
+                                testNames.add(createMethodName(psiClass, method))
+                                initialTestCodes.add(code)
+                            }
+                        }
+                    }
+                } catch (_: Exception) {
+                }
+            }
+            true
+        }
+    }
+
+    private fun createTestSampleClass(imports: String, methodCode: String): String {
+        var normalizedImports = imports
+        if (normalizedImports.isNotBlank()) normalizedImports += "\n\n"
+        return normalizedImports +
+                "public class TestSample {\n" +
+                "   $methodCode\n" +
+                "}"
+    }
+
+    private fun createMethodName(psiClass: PsiClass, method: PsiMethod): String =
+        "<html>${psiClass.qualifiedName}#${method.name}</html>"
 }
